rc1

+++ Sağlık_Ocağı_HTS/Formlar/HastaIslem/YeniSevkForm.cs
@@ -31,6 +31,7 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
             aktifSevkTarihi = DateTime.Now;
             aktifSaat = aktifSevkTarihi.ToString("HH:mm:ss");
             
+            PersonelComboDoldur();
             DrComboDoldur();
             PoliklinikComboDoldur();
             İşlemComboDoldur();
@@ -48,11 +49,12 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
         public YeniSevkForm(sevk s):this()//Düzenleme
         {
             materialSingleLineTextField1.Text = s.sira;
+            materialSingleLineTextField1.Enabled = false;
             ActiveSevk = s;
 
            // ActiveHasta = db.dosya.ToList().First(a =>a.dosyaid == db.sevkler.ToList().First(b =>b.sevktarihi == s.sevktarihi).dosyaID).hasta;
-               
-            ActiveHasta = db.sevk.Where(a => a.sevktarihi == s.sevktarihi).First().dosya.hasta;
+            //var test = db.sevk.ToList().Where(a => a.sevktarihi == s.sevktarihi).FirstOrDefault();
+            ActiveHasta = db.sevk.ToList().Where(a => a.sevktarihi == s.sevktarihi).First().dosya.hasta;
 
             aktifSevkTarihi = s.sevktarihi;
             aktifSaat = aktifSevkTarihi.ToString("HH:mm:ss");
@@ -61,6 +63,7 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
             button2.Text = "Sevki Kaydet";
             DüzenlemeModu = true;
             SevktenPoliComboSec(s);
+            SevktenDrComboSec(s);
         }
 
         #region ComboboxFonksiyonlar Ve sınıfları
@@ -74,11 +77,29 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
                     if (var.poliklinik.poliklinikadi == s.poliklinik)
                     {
                         comboBox1.SelectedIndex = j;
+                        comboBox1.Enabled = false;
                         break;
                     }
                 }
             }
         }
+        void SevktenDrComboSec(sevk s)
+        {
+            for (int j = 0; j < comboBox3.Items.Count; j++)
+            {
+                if (comboBox3.Items[j] is DrComboItem)
+                {
+                    var var = comboBox3.Items[j] as DrComboItem;
+                    if (var.dr.doktorid == s.sevkedendoktorid)
+                    {
+                        comboBox3.SelectedIndex = j;
+                        comboBox3.Enabled = false;
+                        break;
+                    }
+                }
+            }
+        }
+
         void PoliklinikComboDoldur()
         {
             comboBox1.Items.Clear();
@@ -116,6 +137,18 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
                 comboBox3.Items.Add(item);
             }
             comboBox3.SelectedIndex = -1;
+        }
+        void PersonelComboDoldur()
+        {
+            comboBox4.Items.Clear();
+            comboBox4.Items.Add(yeniStr);
+            if (db.personel.Count() == 0)
+                return;
+            foreach (var item in db.personel)
+            {
+                comboBox4.Items.Add(item);
+            }
+            comboBox4.SelectedIndex = -1;
         }
         public class PoliComboItem
         {
@@ -161,7 +194,7 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
 
             foreach (var islemler in islemlers)
             {
-                IslemItem item= new IslemItem(islemler.islem);
+                IslemItem item= new IslemItem(islemler);
                 item.silBtnEvent += IslemItemSilBtnAksiyon;
                 flowLayoutPanel1.Controls.Add(item);
             }
@@ -190,14 +223,26 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
             if (res != DialogResult.Yes)
                 return;
 
-            db= new saglikDBEntities_1();
-            islemler islm = db.islemler.ToList().Where(a => a.islemid == i.islemid).First();
-            db.islemler.Remove(islm);
-            db.SaveChanges();
-            
             IslemItem silinecekitem = flowLayoutPanel1.Controls.Cast<IslemItem>()
                 .Where(a => a.ActiveIslem.islemid == i.islemid).First();
             flowLayoutPanel1.Controls.Remove(silinecekitem);
+
+            if (islemlerList.Contains(silinecekitem.ActiveIslemler))
+            {
+                islemlerList.Remove(silinecekitem.ActiveIslemler);
+            }
+
+            if (DüzenlemeModu)
+            {
+                db= new saglikDBEntities_1();
+                islemler islm = db.islemler.ToList().Where(a => a.islemid == i.islemid).First();
+                db.islemler.Remove(islm);
+                db.SaveChanges();
+            }
+
+          
+            
+            
         }
 
         private void button1_Click(object sender, EventArgs e)
@@ -209,9 +254,9 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
                 MessageBox.Show("Lütfen İşlem Seçiniz!","Hata",MessageBoxButtons.OK,MessageBoxIcon.Warning);
                 return;
             }
-            if (comboBox3.SelectedIndex==-1 || comboBox3.Items[comboBox3.SelectedIndex].ToString()==yeniStr )
+            if (comboBox4.SelectedIndex==-1 || comboBox4.Items[comboBox4.SelectedIndex].ToString()==yeniStr )
             {
-                MessageBox.Show("Lütfen Doktor Seçiniz!","Hata",MessageBoxButtons.OK,MessageBoxIcon.Warning);
+                MessageBox.Show("Lütfen Personel Seçiniz!","Hata",MessageBoxButtons.OK,MessageBoxIcon.Warning);
                 return;
             }
 
@@ -225,20 +270,25 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
             
             islem islem =(comboBox2.SelectedItem as İşlemComboItem).islm;
 
-            IslemItem item= new IslemItem(islem);
-            item.silBtnEvent += IslemItemSilBtnAksiyon;
-            flowLayoutPanel1.Controls.Add(item);
+            
 
             
         
-            doktor dr =(comboBox3.SelectedItem as DrComboItem).dr;
+            //doktor dr =(comboBox3.SelectedItem as DrComboItem).dr;
 
             islemler islemler= new islemler();
             islemler.sevktarihi = aktifSevkTarihi;
-            islemler.doktorid = dr.doktorid;
+            islemler.personelid = (comboBox4.SelectedItem as personel).personelid;
             islemler.miktar = (int)numericUpDown1.Value;
             islemler.islemid = islem.islemid;
+            islemler.personel = (comboBox4.SelectedItem as personel);
+            islemler.islem = islem;
+
             islemlerList.Add(islemler);
+
+            IslemItem item= new IslemItem(islemler);
+            item.silBtnEvent += IslemItemSilBtnAksiyon;
+            flowLayoutPanel1.Controls.Add(item);
         }
 
         #region Yeni işlem ekle Comboboxlar
@@ -274,7 +324,7 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
             if (box.SelectedIndex != -1 && box.Items[box.SelectedIndex].ToString() == yeniStr)
             {
 
-                DoktorEkleForm form = new DoktorEkleForm();
+                DrVePersonelEkleForm form = new DrVePersonelEkleForm(DrVePersonelEkleForm.EklemeTürü.Doktor);
                 if (form.ShowDialog() == DialogResult.OK)
                 {
                     db = new saglikDBEntities_1();
@@ -336,14 +386,14 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
             { 
                
               
-                sevk sevk= db.sevk.Where(a=>ActiveSevk.sevktarihi==a.sevktarihi).FirstOrDefault();
-                sevk.sevktarihi = aktifSevkTarihi;
-                sevk.poliklinik =poliklinik.poliklinikadi;
-                sevk.sira = sıraNo.ToString();
-                db.Entry(sevk).State = EntityState.Modified;
+                //sevk sevk= db.sevk.ToList().Where(a=>ActiveSevk.sevktarihi==a.sevktarihi).FirstOrDefault();
+               
+                ////sevk.poliklinik =poliklinik.poliklinikadi;
+                ////sevk.sira = sıraNo.ToString();
+                //db.Entry(sevk).State = EntityState.Modified;
                 
-                db.SaveChanges();
-                db.Dispose();
+                //db.SaveChanges();
+                //db.Dispose();
             }
            
 
@@ -352,6 +402,12 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
             foreach (var islems in islemlerList)
             {
                 islems.sevktarihi = aktifSevkTarihi;
+
+                //db.Entry(islems.personel).State = EntityState.Unchanged;
+                //db.Entry(islems.islem).State = EntityState.Unchanged;
+                islems.personel = null;
+                islems.islem = null;
+
                 db.islemler.AddOrUpdate(islems);
                 db.SaveChanges();
             }
@@ -359,6 +415,32 @@ namespace Sağlık_Ocağı_HTS.Formlar.HastaIslem
             
             MessageBox.Show($"Sevk başarıyla {((!DüzenlemeModu)?"oluşturuldu":"kaydedildi")}!","Başarılı",MessageBoxButtons.OK,MessageBoxIcon.Information);
             DialogResult = DialogResult.OK;
+
+        }
+
+        private void comboBox4_Format(object sender, ListControlConvertEventArgs e)
+        {
+            if (e.Value==yeniStr)
+            {
+                return;
+            }
+            var personel = e.ListItem as personel;
+            
+            e.Value = personel.birey.ad+" "+personel.birey.soyad;
+        }
+
+        private void comboBox4_SelectionChangeCommitted(object sender, EventArgs e)
+        {
+            ComboBox box = sender as ComboBox;
+            if (box.SelectedIndex != -1 && box.Items[box.SelectedIndex].ToString() == yeniStr)
+            {
+                DrVePersonelEkleForm form = new DrVePersonelEkleForm(DrVePersonelEkleForm.EklemeTürü.Personel);
+                if (form.ShowDialog() == DialogResult.OK)
+                {
+                    db = new saglikDBEntities_1();
+                    PersonelComboDoldur();
+                }
+            }
 
         }
     }
